makefile improvement for starting dev

+++ Makefile
@@ -6,7 +6,7 @@ all: cascade deb
 
 clean:
 	rm -rf cascade
-	ls -t ./pkg/*.deb | sed -e '1,$(KEEP)d' | xargs -d '\n' rm -rf
+	[ -f ./pkg ] && ls -t ./pkg/*.deb | sed -e '1,$(KEEP)d' | xargs -d '\n' rm -rf || true
 	rm -rf ./build
 
 cascade: clean
